@@ -2,14 +2,12 @@
 
 namespace App\Http\Controllers\Auth;
 
-use App\User;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Str;
+use App\Providers\RouteServiceProvider;
+use App\Models\User;
+use Illuminate\Foundation\Auth\RegistersUsers;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\Foundation\Auth\RegistersUsers;
-use App\Invitation_code;
-use Alert;
 
 class RegisterController extends Controller
 {
@@ -31,7 +29,7 @@ class RegisterController extends Controller
      *
      * @var string
      */
-    protected $redirectTo = '/login';
+    protected $redirectTo = RouteServiceProvider::HOME;
 
     /**
      * Create a new controller instance.
@@ -51,60 +49,25 @@ class RegisterController extends Controller
      */
     protected function validator(array $data)
     {
-        $validator =  Validator::make($data, [
-            'first_name' => ['required', 'string', 'max:255'],
-            'last_name' => ['required', 'string', 'max:255'],
+        return Validator::make($data, [
+            'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'password' => ['required', 'string', 'min:6', 'confirmed'],
-            // 'kode_undangan' => ['required','exists:invitation_codes,code']
-            'username' => ['required','string','min:4','max:16','unique:users'],
-            'phone' => ['required','numeric','digits_between:5,15']
+            'password' => ['required', 'string', 'min:8', 'confirmed'],
         ]);
-
-        if($validator->fails()) {
-            return $validator;
-        }
-        // $inv = Invitation_code::where('code',$data['kode_undangan'])->first();
-        // if($inv->remains == 0) {
-        //     Alert::error('Kode undangan sudah dipakai');
-        //     session()->flash('danger','Kode undangan sudah dipakai');
-        //     return redirect()->back();
-        // }else{
-        //     $update = Invitation_code::find($inv->id);
-        //     $update->remains -= 1;
-        //     if($inv->remains == 1) {
-        //         $update->status = "Redeemed";
-        //     }
-        //     $update->save();
-
-
-        Alert::success('Sukses daftar!','Sukses');
-        session()->flash('success','Sukses daftar! Silahkan login');
-            
-        // }
-        return $validator;
     }
 
     /**
      * Create a new user instance after a valid registration.
      *
      * @param  array  $data
-     * @return \App\User
+     * @return \App\Models\User
      */
     protected function create(array $data)
     {
-
         return User::create([
-            'name' => $data['first_name']." ".$data['last_name'],
+            'name' => $data['name'],
             'email' => $data['email'],
-            'username' => $data['username'],
-            'phone' => $data['phone'],
             'password' => Hash::make($data['password']),
-            'balance' => 0,
-            'level' => 'Member',
-            'status' => 'Active',
-            'api_key' => Hash::make(Str::random(5)),
-            'uplink' => 'Server',
         ]);
     }
 }
